test: consolidate mock_ws_client fixture into tests/platform/conftest.py
